// web/src/lib/types.ts
type Images = {
  jpg: SizeImages;
  webp: SizeImages;
};
// ...
type Character = {
  mal_id: number;
  url: string;
  images: CharacterImages;
  name: string;
};

type CharacterImages = {
  jpg: Jpg;
  webp: Webp;
};

type Jpg = {
  image_url: string;
};
// ...
export interface Location<State = IDLocation> extends Path {
  state: State;
  key: string;
}

type IDLocation = {
  id: number;
};

type Path = {
  pathname: string;
  search: string;
  hash: string;
};

//type Actor Voice

export type ActorVoiceData = {
  data: ActorVoice;
};

export type ActorVoice = {
  mal_id: number;
  url: string;
  website_url: null;
  images: ActorImages;
  name: string;
  given_name: string;
  family_name: string;
  alternate_names: any[];
  birthday: string;
  favorites: number;
  about: string;
  anime: AnimeElement[];
  manga: any[];
  voices: Voice[];
};

type AnimeElement = {
  position: string;
  anime: AnimePositions;
};

type AnimePositions = {
  mal_id: number;
  url: string;
  images: Images;
  title: string;
};

type ActorImages = {
  jpg: Jpg;
};

export type Voice = {
  role: string;
  anime: AnimePositions;
  character: Character;
};